refactor: extract MetricsChartRenderer, shared by both carousel charts
ADFA-5489 gave the metrics carousel a second chart, and with it a second
copy of the chart setup: the two renderers had a byte-identical
configure() apart from the value formatter, and a byte-identical block
in rebuild() applying theme colours and redrawing.

ADFA-5486 adds x-axis labels, zoom, event annotations and snapshot
export to "the line chart", written when there was only one. All four
belong on both charts, and duplicated setup is how they end up on one.
This puts the common behaviour in one place before that work starts.

MetricsChartRenderer holds the attach/detach lifecycle -- including
detachIfAttached, which a recycling carousel page needs -- the shared
axis and gesture configuration, and the data/redraw helpers. Subclasses
override configure() to add what is theirs (the memory chart's MB
formatter; the network chart's byte formatter and per-decade
granularity) and call through.

Behaviour-neutral: no configuration value changed, only where it lives.
The existing renderer tests are the evidence, and both charts were
compared on device against the previous build.

Verified: :app:testV8DebugUnitTest, 51 tests green across app ui/utils;
both carousel pages rendered on a Pixel 6 Pro (arm64, v8 debug),
including the network chart under a live Gradle sync.

ADFA-5486
